requirements.txt
annotated-types==0.7.0
base58==2.1.1
based58==0.1.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ciso8601==2.3.2
click==8.2.1
filelock==3.18.0
grpcio==1.73.0
grpcio-tools==1.73.0
idna==3.10
importlib_metadata==8.7.0
loguru==0.7.3
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.1
orjson==3.10.18
panther @ git+ssh://git@github.com/Chindada/panther.git/@a3f95ae6f3ab21a3adc230aab6a2976375b0b292
prometheus_client==0.22.1
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyNaCl==1.5.0
pyrsca==0.1.2
pysolace==0.9.52
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
sentry-sdk==2.30.0
setuptools==80.9.0
shellingham==1.5.4
shioaji==1.2.6
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
xxhash==3.5.0
zipp==3.23.0
